requirements.txt
pytest==6.0.1
loguru==0.5.1
bcrypt==3.2.0
passlib==1.7.4
fastapi==0.54.1
uvicorn==0.11.5
aiomysql==0.0.20
python-jose==3.2.0
cryptography==3.1.1
python-dotenv==0.13.0
tortoise-orm==0.16.17
python-multipart==0.0.5